+++ fe/src/main/java/org/apache/doris/persist/EditLog.java
@@ -212,13 +212,13 @@ public class EditLog {
                     ModifyPartitionInfo info = (ModifyPartitionInfo) journal.getData();
                     LOG.info("Begin to unprotect modify partition. db = " + info.getDbId()
                             + " table = " + info.getTableId() + " partitionId = " + info.getPartitionId());
-                    catalog.replayModifyPartition(info);
+                    catalog.getAlterInstance().replayModifyPartition(info);
                     break;
                 }
                 case OperationType.OP_BATCH_MODIFY_PARTITION: {
                     BatchModifyPartitionsInfo info = (BatchModifyPartitionsInfo) journal.getData();
                     for(ModifyPartitionInfo modifyPartitionInfo : info.getModifyPartitionInfos()) {
-                        catalog.replayModifyPartition(modifyPartitionInfo);
+                        catalog.getAlterInstance().replayModifyPartition(modifyPartitionInfo);
                     }
                     break;
                 }
